Move repeat block simulation into repeat module
